Addressed lighting issues on WallGenExtrusion's meshes
Because each surface was sharing vertices Unity was treating them as a contiguous surface, which meant the lighting was behaving badly, attempting to render a 90 degree edge as a smooth surface.  Instead, each face needs to have its own vertices - we can get away with the triangles on a face sharing vertices because they are planar.